requirements.txt
# Essential dependencies for your Streamlit app
streamlit==1.48.0
pandas==2.2.0
requests==2.32.3
python-dotenv==1.1.0
openai==1.68.2
plotly==6.1.2
python-dateutil==2.9.0.post0
